% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{bayesfit}
\alias{bayesfit}
\alias{plot.bayesfit}
\title{Traces from the MCMC output}
\usage{
bayesfit()

\method{plot}{bayesfit}(x, which = 1:3, ...)
}
\arguments{
\item{x}{an arbitrary \R object.}

\item{which}{a vector with values in {1,2,3} where 1 is to plot residual
density functions, 2 is for residual distribution functions and 3 is for
a contour plot of the posterior distribution function of (alpha, beta).}
}
\value{
An object of class "bayesfit" with MCMC traces for:
\item{a, b }{Heffernan--Tawn model parameters (matrices).}
\item{sd, mean }{means and standard deviations of the residual mixture components (3-dimensional arrays).}
\item{w }{weights of the mixture components (matrix).}
\item{prec }{precision parameter of the Dirichlet process (vector).}
\item{ci }{auxiliary variable; components' indices for each observation (matrix).}
\item{noo }{number of observation in each mixture component (matrix).
\item{noc }{number of non-empty components in the mixture, i.e., for which at least one data points has a component index pointing to it.}
\item{prop.sd }{standard deviations for the proposal distributions of \code{a} and \code{b} (3-dimensional array).}
And \code{len}, the length of the traces, i.e., the number of iterations
saved after burning and thinning; \code{nlag} stores the number of lags
present in the data.
}
\description{
Test or show objects of class "bayesfit".
}
\details{
Default plot shows samples of residual densities (\code{which==1}), residual
distribution with credible interval (\eqn{5\%} and \eqn{95\%} posterior
quantile; \code{which==2}), and joint posterior distribution of \eqn{\alpha}
and \eqn{\beta} (\code{which==3}) for each lag successively. \code{which}
can be any composition of 1,2 and 3.
}
\seealso{
\code{\link[=bayesparams]{bayesparams()}}, \code{\link[=stepfit]{stepfit()}}
}
