% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empfit.R
\name{block.bootstrap.sample}
\alias{block.bootstrap.sample}
\title{Compute a block bootstrap sample from a time series}
\usage{
block.bootstrap.sample(ts, block.length, R)
}
\arguments{
\item{ts}{vector of reals, time series which to estimate theta(x,m) for}

\item{R}{integer>0, nbr of repetitions used for the block-bootstrap CI}

\item{block.length:}{integer>0, used for the block-bootstrap CI}
}
\value{
matrix, each column is a block-bootstrapped time series
}
\description{
Called by \code{\link[=thetaruns]{thetaruns()}}.
}
\keyword{internal}
