% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et2fit.R
\name{blockboot}
\alias{blockboot}
\title{Basis method to block bootstrap a time series}
\usage{
blockboot(ts, block_length, R)
}
\arguments{
\item{block_length}{integer>0, length of blocks in block-bootstrap.}

\item{R}{integer>0, number of bootstrap samples to compute the confidence
intervals.}
}
\value{
A matrix, \link{R} columns containing the block bootstrap samples of \link{ts}
}
\description{
Called by \code{\link[=thboot]{thboot()}}, not exposed to the user.
}
\keyword{internal}
