% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empfit.R
\name{compute.runs}
\alias{compute.runs}
\title{Compute the extremal index using the runs method}
\usage{
compute.runs(data, mesh)
}
\arguments{
\item{data}{matrix, lags matrix of a time series}

\item{mesh}{vector, x in theta(x,m) on Laplace scale}
}
\value{
list, estimates of theta across \link{mesh} - nbr of exceedances
across \link{mesh}
}
\description{
Called by \code{\link[=thetaruns]{thetaruns()}}.
}
\keyword{internal}
