% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{conditions.verify}
\alias{conditions.verify}
\title{Bounds on (alpha,beta)}
\usage{
conditions.verify(a, b, p, v = -log(2 * (1 - 0.99999999)), data)
}
\arguments{
\item{a}{scalar, alpha parameter, in \link{-1,1}}

\item{b}{scalar, beta parameter, in \link{0,1}}

\item{p}{scalar, probability (to compute quantiles of residual distribution), in \link{0,1}}

\item{v}{scalar, threshold, recommended as beyond the maximum observation}

\item{data}{bivariate vector, (X,Y) with Y | X>u}

\item{boolean, }{does the couple (a,b) satisfy the conditions?}
}
\value{
A boolean, TRUE if the (alpha, beta) pair is valid, FALSE otherwise.
}
\description{
Mainly called by \code{\link[=conditions.bounds]{conditions.bounds()}} to check whether a given (alpha, beta)
pair lies within the Keef et al. boundaries.
}
\keyword{internal}
