% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{depmeasure}
\alias{depmeasure}
\alias{is.depmeasure}
\alias{plot.depmeasure}
\title{Dependence measures estimates}
\usage{
depmeasure(type = c("theta", "chi", "steptheta", "runs"))

is.depmeasure(x)

\method{plot}{depmeasure}(x, ...)
}
\arguments{
\item{type}{one of "theta", "chi", "steptheta" or "runs".}

\item{x}{an arbitrary \R object.}

\item{...}{other parameters to be passed through to printing or plotting
functions.}
}
\value{
An object of class "depmeasure" whose exact structure depends on
its type. All types contain the following elements,
\item{probs }{probability levels at which the dependence measure was estimated.}
\item{levels }{corresponding quantile levels on the data's original scale.}
\item{nlag }{number of lags.}

Furthermore, if \code{type} is "theta" or "chi",
\item{fit }{an object of class "bayesfit".}
\item{distr }{a matrix with samples from the posterior distribution, with \code{length(levels)} rows and \code{S} columns (see \code{\link[=thetafit]{thetafit()}}).}
\item{theta,chi }{a matrix with the posterior estimates of the respective dependence measure at different thresholds \code{levels} (rows) and coverage intervals (named columns, first two for the mean and median).}

If \code{type} is "steptheta",
\item{fit }{an object of class "stepfit".}
\item{theta }{a matrix with the estimate for theta at different thresholds \code{levels} (rows) and confidence levels (columns).}

If \code{type} is "runs",
\item{theta }{a matrix with the (empirical) runs estimator estimate and confidence intervals (columns) at the given \code{levels} (rows).}
\item{nbr_exc }{the number of exceedances corresponding to the \code{levels}.}
}
\description{
Test or show objects of class "depmeasure", a structure for all dependence
measures covered in this package.
}
