% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{dicho.b.bounds}
\alias{dicho.b.bounds}
\alias{dicho.a.bounds}
\title{Determine exact bound for alpha or beta}
\usage{
dicho.b.bounds(i, o, fix.par, data, eps, v = -log(2 * (1 - 0.99999999)))

dicho.a.bounds(i, o, fix.par, data, eps, v = -log(2 * (1 - 0.99999999)))
}
\arguments{
\item{i}{current inner bound}

\item{o}{current outer bound}

\item{fix.par}{scalar, either value of alpha or of beta}

\item{data}{bivariate vector, (X,Y) with Y | X>u}

\item{eps}{scalar, tolerance, typically 0.001}
}
\value{
Scalar, the bound found.
}
\description{
Pair of functions called by \code{\link[=conditions.bounds]{conditions.bounds()}} to get more accurate
bounds by dichotomy.
}
\keyword{internal}
