% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{format_ts}
\alias{format_ts}
\title{Pre-process time series to feed Heffernan--Tawn model}
\usage{
format_ts(
  ts,
  u_mar,
  u_dep,
  method = c("mle", "mom", "pwm"),
  nlag,
  lapl = FALSE
)
}
\arguments{
\item{ts}{vector of reals, (stationary!) time series on an arbitrary scale.}

\item{u_mar}{scalar, marginal (GPD) threshold given as a probability.}

\item{u_dep}{scalar, conditional threshold given as a probability.}

\item{method}{string, "mle" for max likelihood or "mom" for the method of
moments or "pwm" for proba weighted moments}

\item{nlag}{integer, number of lags to be considered.}

\item{lapl}{logical, is \link{ts} on Laplace scale already?}
}
\value{
A matrix with \link{nlag}+1 columns on the Laplace scale, with first
column above threshold.
}
\description{
Called by \code{\link[=etfit]{etfit()}} and \code{\link[=et2fit]{et2fit()}}, not exposed to the user.
}
\keyword{internal}
