% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpd.mle}
\alias{gpd.mle}
\title{Maximum likelihood estimation for the GPD}
\usage{
gpd.mle(ts, u, hessian = FALSE)
}
\arguments{
\item{ts}{vector of reals, stationary time series.}

\item{u}{scalar, threshold given as a quantile.}

\item{hessian}{boolean, TRUE means computing the hessian matrix of the
likelihood.}
}
\value{
A list, threshold - MLEs - log-likelihood - standard deviations -
covariance matrix.
}
\description{
Called by \code{\link[=scale.ts]{scale.ts()}}, not exposed to the user.
}
\keyword{internal}
