% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpd.mom}
\alias{gpd.mom}
\alias{gpd.pwm}
\title{Method of moments and probability weighted moments to fit the GPD}
\usage{
gpd.mom(ts, u)

gpd.pwm(ts, u)
}
\arguments{
\item{ts}{vector of reals, stationary time series.}

\item{u}{scalar, threshold given as a quantile.}
}
\value{
A list, threshold - MOM estimates.
}
\description{
Called by \code{\link[=scale.ts]{scale.ts()}}, not exposed to the user.
}
\keyword{internal}
