% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpd}
\alias{gpd}
\title{GPD fit wrapper}
\usage{
gpd(ts, u, method)
}
\arguments{
\item{ts}{vector of reals, stationary time series.}

\item{u}{scalar, threshold given as a quantile.}
}
\description{
GPD fit wrapper
}
\keyword{internal}
