% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpd_mom}
\alias{gpd_mom}
\alias{gpd_pwm}
\title{Method of moments and probability weighted moments to fit the GPD}
\usage{
gpd_mom(ts, u)

gpd_pwm(ts, u)
}
\arguments{
\item{ts}{vector of reals, stationary time series.}

\item{u}{scalar, threshold given as a quantile.}
}
\value{
A list, threshold - MOM estimates.
}
\description{
Called by \code{\link[=scale_ts]{scale_ts()}}, not exposed to the user.
}
\keyword{internal}
