% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht2fit.R
\name{ht2step.2d}
\alias{ht2step.2d}
\title{Heffernan--Tawn model stepwise fit in two dimensions}
\usage{
ht2step.2d(data, conditions = TRUE)
}
\arguments{
\item{data}{2-column matrix, first column is \eqn{X>u} (Laplace) - second
column is \eqn{Y|X>u}}

\item{conditions}{boolean, TRUE means alpha and beta must satisfy AD/AI constraints}
}
\value{
list, MLE for alpha - beta - z hat - standard deviations of alpha and beta
}
\description{
Called by \code{\link[=ht2step]{ht2step()}}
}
\keyword{internal}
