% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht2fit.R
\name{ht2step}
\alias{ht2step}
\title{Heffernan--Tawn model stepwise fit in the general case}
\usage{
ht2step(data, conditions = TRUE)
}
\arguments{
\item{data}{matrix of reals, first column is X>u (Laplace) - other columns are |X>u}

\item{conditions}{boolean, TRUE means alpha and beta must satisfy AD/AI constraints}
}
\value{
An object of class \code{\link[=stepfit]{stepfit()}}.
}
\description{
Called by \code{\link[=th2fit]{th2fit()}}
}
\keyword{internal}
