% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_int}
\alias{is_int}
\title{A loose integer validation function}
\usage{
is_int(x)
}
\arguments{
\item{x}{an arbitrary \R object.}
}
\description{
A loose integer validation function
}
\keyword{internal}
