% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{lags.matrix}
\alias{lags.matrix}
\title{Build matrix of lagged observations}
\usage{
lags.matrix(ts, nlag)
}
\arguments{
\item{ts}{vector of reals, (stationary!) time series.}

\item{nlag}{scalar, number of lags of the output lagged matrix
(=> number of columns == \code{nlag}+1).}
}
\value{
A matrix of lagged versions of \code{ts} in its columns.
}
\description{
Called by \code{\link[=format.ts]{format.ts()}} and \code{\link[=thetaruns]{thetaruns()}}, not exposed to the user.
}
\keyword{internal}
