% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{nllh.gpd.0}
\alias{nllh.gpd.0}
\title{Negtive log-likelihood for the GPD with null shape}
\usage{
nllh.gpd.0(par, u, x)
}
\arguments{
\item{par}{scalar, scale (shape is assumed 0).}

\item{u}{scalar, threshold given as a quantile.}

\item{x}{vector of reals, data on which to fit the GPD.}
}
\value{
A scalar, value of the negative log-likelihood of the GPD
}
\description{
Called by \code{\link[=gpd.mle]{gpd.mle()}} when \code{shape.null.hyp==TRUE}, not exposed to the user.
}
\keyword{internal}
