% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{nllh_gpd}
\alias{nllh_gpd}
\alias{grad_gpd}
\title{Negative log-likelihood and its gradient for the GPD}
\usage{
nllh_gpd(par, u, x)

grad_gpd(par, u, x)
}
\arguments{
\item{par}{vector of 2 reals, scale-shape.}

\item{u}{scalar, threshold given as a quantile.}

\item{x}{vector of reals, data on which to fit the GPD.}
}
\value{
A scalar, value of the negative log-likelihood of the GPD.
}
\description{
Called by \code{\link[=gpd_mle]{gpd_mle()}}, not exposed to the user.
}
\keyword{internal}
