% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht2fit.R
\name{nllh_ht}
\alias{nllh_ht}
\alias{grad_ht}
\title{Negative log-likelihood for the Heffernan--Tawn model}
\usage{
nllh_ht(par, data, conditions)

grad_ht(par, data, conditions)
}
\arguments{
\item{par}{vector of reals, alpha-beta-mu-sigma^2}

\item{data}{2-column matrix, first column is X>u - second column is Y|X>u}

\item{conditions}{boolean, TRUE means alpha and beta must satisfy AD/AI constraints}
}
\value{
scalar, value of the negative log-likelihood of H+T
}
\description{
Called by \code{\link[=ht2step_2d]{ht2step_2d()}}
}
\keyword{internal}
