% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht2distr.R
\name{p.res2}
\alias{p.res2}
\alias{q.res2}
\title{The conditional tail model residual distribution (stepwise)}
\usage{
p.res2(res, sorted.res)

q.res2(p, a, b, data)
}
\arguments{
\item{res}{vector of reals, quantiles at which to compute the distribution
function of the residuals.}

\item{sorted.res}{vector, empirical distribution function.}

\item{p}{scalar, probability (to compute quantiles of residual distribution),
in \link{0,1}.}

\item{a}{scalar, alpha parameter, in \link{-1,1}.}

\item{b}{scalar, beta parameter, in \link{0,1}.}

\item{data}{bivariate vector, \eqn{(X,Y)} with \eqn{Y | X>u}.}
}
\value{
For \code{p.res2}: a vector of the same length as \code{res}, distribution of
the residuals evaluated in \code{res}. For \code{q.res2}: a quantile of the residual
distribution.
}
\description{
The \code{p.res2} method is called by \code{\link[=th2est]{th2est()}}, the \code{q.res2} method is called
by \code{\link[=verifies.conditions]{verifies.conditions()}} in order to get quantiles of the empirical
distribution function of the residual distribution given
\eqn{(\alpha,\beta)}. Not exposed to the user.
}
\keyword{internal}
