% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htdistr.R
\name{p_res}
\alias{p_res}
\alias{r_res}
\title{The conditional tail model residual distribution (Bayesian)}
\usage{
p_res(s, z, w, mu, sig)

r_res(R, S, nlag, w, mu, sig)
}
\arguments{
\item{s}{integer, ranges across sweeps.}

\item{z}{array of reals, trace of residuals.}

\item{w}{matrix of reals, trace of the weights of components --- rows sum
to 1.}

\item{mu, sig}{arrays of reals, means and standard deviations of the stored
mixture components output from \code{\link[=depfit]{depfit()}} with dimensions
\emph{nbr of sweeps} x \emph{number of components} x \emph{nbr of lags}.}

\item{R}{integer>0, number of samples from a mixture distribution sampled at
a given iteration.}

\item{S}{integer>0, number of posterior iterations used for the estimation.}

\item{nlag}{integer>0, number of lags in the original data.}
}
\value{
For \code{p_res}: a matrix of reals, evaluation of the Heffernan--Tawn
residual distribution function. For \code{r_res}: a matrix of reals, samples
of the residual distribution across a sample of iterations.
}
\description{
Distribution function and random generation for the residual distribution of
the Heffernan--Tawn conditional tail model. Intended for the Bayesian
semiparametric approach. The \code{p_res} method is called by
\code{\link[=etfit]{etfit()}} for Monte-Carlo integration, the \code{r_res} method is called by
\code{\link[=etfit]{etfit()}} for the proportion method. Not exposed to the user.
}
\keyword{internal}
