% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{scale.to.original}
\alias{scale.to.original}
\title{Use fitted transform to scale probabilities back to the original scale}
\usage{
\method{scale}{to.original}(p, ts, u, gpd.pars)
}
\arguments{
\item{p}{(vector of) reals, (a range of) probabilities which to transform to
original \link{ts} scale.}

\item{ts}{vector of reals, time series in original scale.}

\item{u}{scalar, threshold given as a probability.}

\item{gpd.pars}{vector of 2 reals, scale-shape parameters of the GPD fitted
to \link{ts}.}
}
\value{
A vector of the same length as \link{p} on \link{ts} scale.
}
\description{
Called by \code{\link[=thetafit]{thetafit()}}, \code{\link[=chifit]{chifit()}}, \code{\link[=theta2fit]{theta2fit()}}
}
\keyword{internal}
