% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{scale.ts}
\alias{scale.ts}
\title{Marginal transformation to Laplace scale}
\usage{
\method{scale}{ts}(ts, u, method = c("mle", "mom", "pwm"))
}
\arguments{
\item{ts}{vector of reals, (stationary!) time series.}

\item{u}{scalar, threshold given as a probability.}

\item{method}{string, "mle" for max likelihood or "mom" for the method of
moments or "pwm" for proba weighted moments.}
}
\value{
A list, original time series with different marginal distributions +
GPD parameters.
}
\description{
Called by \code{\link[=format.ts]{format.ts()}}, not exposed to the user.
}
\keyword{internal}
