% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{scale_to_margin}
\alias{scale_to_margin}
\title{Use fitted transform to scale data to different scales}
\usage{
scale_to_margin(x, ts, u, gpd_pars)
}
\arguments{
\item{x}{(vector of) reals, (a range of) observations for which change of
scale is needed.}

\item{ts}{vector of reals, time series in original scale.}

\item{u}{scalar, threshold given as a probability.}

\item{gpd_pars}{vector of 3 reals, scale-shape-probability of exceeding u.}
}
\value{
A list, original observation(s) scaled to different marginal
distributions: Uniform, Gumbel, Fréchet and Laplace.
}
\description{
Currently not used, not exposed to the user.
}
\keyword{internal}
