% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{scale_to_original}
\alias{scale_to_original}
\title{Use fitted transform to scale probabilities back to the original scale}
\usage{
scale_to_original(p, ts, u, gpd_pars)
}
\arguments{
\item{p}{(vector of) reals, (a range of) probabilities which to transform to
original \code{ts} scale.}

\item{ts}{vector of reals, time series in original scale.}

\item{u}{scalar, threshold given as a probability.}

\item{gpd_pars}{vector of 2 reals, scale-shape parameters of the GPD fitted
to \code{ts}.}
}
\value{
A vector of the same length as \code{p} on \code{ts} scale.
}
\description{
Called by \code{\link[=thetafit]{thetafit()}}, \code{\link[=chifit]{chifit()}}, \code{\link[=theta2fit]{theta2fit()}}
}
\keyword{internal}
