% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et2fit.R
\name{th2est}
\alias{th2est}
\title{Estimate the extremal index from a Heffernan--Tawn model fit}
\usage{
th2est(fit, sim_U, sim_Z, mesh, nlag, method)
}
\arguments{
\item{fit}{output from \code{\link[=ht2step]{ht2step()}}, an object of class \code{\link[=stepfit]{stepfit()}}.}

\item{sim_U}{vector of reals, \code{R} uniform samples (see \code{\link[=ht2step]{ht2step()}}).}

\item{sim_Z}{matrix of reals, \code{R} multivariate samples from the
Heffernan--Tawn residual distribution.}

\item{mesh}{vector of reals, quantile values at which to evaluate
\eqn{\theta} (uniform scale).}

\item{nlag}{number of lags \eqn{m} to consider}

\item{method}{(vector of) string(s), either "prop" or "MCi" for the method
of proportions (empirical approach) or Monte-Carlo integration.}
}
\value{
Estimate of \eqn{\theta(x,m)} at each value of \eqn{x} given by
\code{mesh}.
}
\description{
Simulate residuals from the fitted Heffernan--Tawn model and build (X,Y)
pairs with \eqn{X>u} to derive an estimate of \eqn{\theta(x,m)}.
Called by \code{\link[=theta2fit]{theta2fit()}}, not exposed to the user.
}
\keyword{internal}
