% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et2fit.R
\name{thboot}
\alias{thboot}
\title{Bootstrap confidence interval for the extremal index}
\usage{
thboot(par, block_length, R, method, levels)
}
\arguments{
\item{par}{list, arguments to feed \code{\link[=theta2fit]{theta2fit()}}.}

\item{block_length}{integer>0, length of blocks in block-bootstrap.}

\item{R}{integer>0, number of bootstrap samples to compute the confidence
intervals.}

\item{method}{string, either "prop" or "MCi".}

\item{levels}{vector of probabilities, confidence levels, quantiles of
bootstrap distribution to be output.}
}
\value{
A matrix where columns are quantiles and rows are x-levels of
\eqn{\theta(x,m)}.
}
\description{
Compute bootstrap confidence intervals for the extremal index estimated from
the stepwise fit of the conditional tail model. Called by \code{\link[=th2est]{th2est()}}, not
exposed to the user.
}
\keyword{internal}
