% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margtransf.R
\name{threshold_margin}
\alias{threshold_margin}
\title{Pre-process matrix of lagged data for Heffernan--Tawn fit}
\usage{
threshold_margin(data, u = 0.95, quantile = FALSE)
}
\arguments{
\item{data}{matrix of reals, replicates in rows - variables in columns.}

\item{u}{scalar, threshold given as a probability or quantile (cf. quantile).}

\item{quantile}{boolean, \code{TRUE} means \code{u} is given as a quantile --- \code{FALSE}
means as a probability.}
}
\value{
A matrix of reals, only rows for which the first column of
observations exceed \code{u} are kept.
}
\description{
Called by \code{\link[=format_ts]{format_ts()}}, not exposed to the user.
}
\keyword{internal}
