% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{toc}
\alias{toc}
\title{Timing function}
\usage{
toc(silent = FALSE)
}
\arguments{
\item{silent}{boolean, should the function be verbose (\code{FALSE}, default), or
only return a time value invisibly?}
}
\value{
A time difference (invisibly).#'
}
\description{
Wrapper around \code{\link[tictoc:tic]{tictoc::tic()}} and \code{\link[tictoc:tic]{tictoc::toc()}}
}
\keyword{internal}
