% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htfit.R
\name{depfit}
\alias{depfit}
\title{Conditional tail model Bayesian fit}
\usage{
depfit(
  ts,
  u_mar = 0,
  u_dep = u_mar,
  lapl = FALSE,
  method_mar = c("mle", "mom", "pwm"),
  nlag = 1,
  par = bayesparams(),
  submodel = c("fom", "none", "ugm"),
  u.mar = deprecated(),
  u.dep = deprecated(),
  method.mar = deprecated()
)
}
\arguments{
\item{ts}{numeric vector; time series to be fitted.}

\item{u_mar}{marginal threshold; used when transforming the time series to
the Laplace scale.}

\item{u_dep}{dependence threshold; level above which the dependence is
modelled. \code{u_dep} can be lower than \code{u_mar}.}

\item{lapl}{logical; is \code{ts} on the Laplace scale already? The default
(FALSE) assumes unknown marginal distribution.}

\item{method_mar}{a character string defining the method used to estimate the
marginal GPD; either \code{"mle"} for maximum likelihood or \code{"mom"}
for method of moments or \code{"pwm"} for probability weighted moments.
Defaults to \code{"mle"}.}

\item{nlag}{integer; number of lags to be considered when modelling the
dependence in time.}

\item{par}{an object of class "bayesparams".}

\item{submodel}{a character string; "fom" for \emph{first order Markov},
"none" for \emph{no particular time structure}, or "ugm" for
\emph{univariate Gaussian mixture} (see Details).}

\item{u.mar}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use \code{u_mar} instead.}

\item{u.dep}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use \code{u_dep} instead.}

\item{method.mar}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} use \code{method_mar}
instead.}
}
\value{
An object of class \code{\link[=bayesfit]{bayesfit()}}.
}
\description{
Bayesian semiparametrics are used to fit the Eastoe--Tawn model to a time
series. Options are available to impose a structure in time on the model.
}
\details{
\code{submodel} can be \code{"fom"} to impose a first order Markov structure
on the model parameters \eqn{\alpha_j} and \eqn{\beta_j}
(see \code{\link[=thetafit]{thetafit()}} for more details); it can take the value \code{"none"} to
impose no particular structure in time; it can also be \code{"ugm"} which
can be applied to density estimation, as it corresponds to setting
\eqn{\alpha=\beta=0} (see examples).#'
}
\examples{
## generate data from an AR(1)
## with Gaussian marginal distribution
n   <- 10000
dep <- 0.5
ar    <- numeric(n)
ar[1] <- rnorm(1)
for(i in 2:n)
  ar[i] <- rnorm(1, mean=dep*ar[i-1], sd=1-dep^2)

## rescale the margin
ar <- qlapl(pnorm(ar))

## fit the data
params <- bayesparams()
params$maxit <- 100# bigger numbers would be
params$burn  <- 10 # more sensible...
params$thin  <- 4
fit <- depfit(ts=ar, u_mar=0.95, u_dep=0.98, par=params)

########
## density estimation with submodel=="ugm"
data <- MASS::galaxies/1e3
dens <- depfit(ts=data, par=params, submodel="ugm")
}
\seealso{
\code{\link[=thetafit]{thetafit()}}, \code{\link[=chifit]{chifit()}}
}
