% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htfit.R
\name{htfit}
\alias{htfit}
\title{Raw method for the semiparametric Bayesian fit}
\usage{
htfit(
  data,
  prop.a,
  prop.b,
  prior.mu = c(0, 10),
  prior.nu = c(2, 1/2),
  prior.eta = c(4, 1),
  trunc = 100,
  comp.saved = 10,
  maxit = 10000,
  burn = 2000,
  thin = 4,
  adapt = 2000,
  batch.size = 125,
  start.ab = c("guesstimate", "prior"),
  mode = 1,
  submodel = c("fom", "none", "ugm")
)
}
\arguments{
\item{data}{matrix, first column is > threshold -- other columns are lagged
versions of the first one.}

\item{submodel}{a character string; "fom" for \emph{first order Markov},
"none" for \emph{no particular time structure}, or "ugm" for
\emph{univariate Gaussian mixture} (see Details).}
}
\value{
An object of class \code{\link[=bayesfit]{bayesfit()}}.
}
\description{
Wrapper function for \code{C_et_interface()}. Called by \code{\link[=depfit]{depfit()}}, not exposed
to the user.
}
\keyword{internal}
