% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{stepfit}
\alias{stepfit}
\alias{plot.stepfit}
\title{Estimates from the stepwise fit}
\usage{
stepfit()

\method{plot}{stepfit}(x, ...)
}
\arguments{
\item{x}{an arbitrary \R object.}
}
\value{
An object of class "stepfit" with the following elements:
\item{a, b }{Heffernan--Tawn model parameters of length \code{nlag}; \eqn{\alpha} controls the conditional extremal expectation, while \eqn{\beta} controls the conditional extremal expectation and variance.}
\item{res }{matrix of fitted residuals with \code{nlag} columns.}
\item{pars.se }{2-column matrix of estimated standard errors for \code{a} (first column) and \code{b} (second column), given by the hessian matrix of the likelihood function used in the first step of the inference procedure.}
\item{nlag }{number of lags.}
}
\description{
Create, test or show objects of class "stepfit".
}
\seealso{
\code{\link[=bayesfit]{bayesfit()}}, \code{\link[=depmeasure]{depmeasure()}}
}
