% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsxtreme-package.R
\docType{package}
\name{tsxtreme-package}
\alias{tsxtreme}
\alias{tsxtreme-package}
\title{tsxtreme: Bayesian Modelling of Extremal Dependence in Time Series}
\description{
Characterisation of the extremal dependence structure of time series, avoiding pre-processing and filtering as done typically with peaks-over-threshold methods. It uses the conditional approach of Heffernan and Tawn (2004) \doi{10.1111/j.1467-9868.2004.02050.x} which is very flexible in terms of extremal and asymptotic dependence structures, and Bayesian methods improve efficiency and allow for deriving measures of uncertainty. For example, the package contains a routine to estimate the extremal index, a measure related to the size of clusters of extreme values in time, and this same routine also gives access to the raw samples from the extremal index posterior distribution.
}
\details{
The Heffernan--Tawn conditional formulation for a stationary time series
\eqn{(X_t)} with Laplace marginal distribution states that for a large enough
threshold \eqn{u} there exist scale parameters \eqn{-1 \le\alpha_j\le 1} and
\eqn{0 \le \beta_j \le 1} such that
\deqn{Pr\left(\frac{X_j-\alpha_j X_0}{(X_j)^{\beta_j}} < z_j, j=1,\ldots,m \mid X_0 > u\right) = H(z_1,\ldots,z_m),}{%
 Pr{(X_j-\alpha_j * X_0)/(X_j)^{\beta_j} < z_j, j=1,\ldots,m | X_0 > u} = H(z_1,\ldots,z_m)}
with \eqn{H} non-degenerate; the equality holds exactly only when \eqn{u}
tends to infinity.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} As of version 0.4.0, all arguments that
included a \code{.} to separate a compound name were renamed to use
\strong{snake_case}, i.e., a \verb{_} instead. For example, the elements of a
\code{bayesparams} object are now \code{prop_a}, \code{prop_b} and so on. The older notation
is still accepted in this version and you will see deprecation warnings; a
later version shall not accept the dot-notation anymore.

There are mainly 3 functions provided by this package, which allow estimation
of extremal dependence measures and fitting the Heffernan--Tawn model using
Dirichlet processes.

\code{\link[=depfit]{depfit()}} fits the Heffernan--Tawn model using a Bayesian
semi-parametric approach.

\code{\link[=thetafit]{thetafit()}} computes posterior samples of the threshold-based
index of Ledford and Tawn (2003) based on inference in \code{\link[=depfit]{depfit()}}.

\code{\link[=chifit]{chifit()}} computes posterior samples of the extremal measure of
dependence of Coles, Heffernan and Tawn (1999) at any extremal level.

Some corresponding functions using the stepwise approach of
Heffernan and Tawn (2004) are also part of the package, namely
\code{\link[=dep2fit]{dep2fit()}} and \code{\link[=theta2fit]{theta2fit()}}.

The empirical estimation of the extremal index can be done using
\code{\link[=thetaruns]{thetaruns()}} and some basic functions handling the Laplace
distribution are also available in \code{\link[=dlapl]{dlapl()}}.
}
\references{
Coles, S., Heffernan, J. E. and Tawn, J. A. (1999) Dependence measures for
extreme value analyses. \emph{Extremes}, \strong{2}, 339--365.

Davison, A. C. and Smith, R. L. (1990) Models for exceedances over high
thresholds. \emph{Journal of the Royal Statistical Society Series B},
\strong{52}, 393--442.

Heffernan, J. E. and Tawn, J. A. (2004) A conditional approach for
multivariate extreme values.
\emph{Journal of the Royal Statistical Society Series B}, \strong{66}, 497--546.

Ledford, W. A. and Tawn, J. A. (2003) Diagnostics for dependence within
time series extremes. \emph{Journal of the Royal Statistical Society Series B},
\strong{65}, 521--543.

Lugrin, T., Davison, A. C. and Tawn, J. A. (2016) Bayesian uncertainty
management in temporal dependence of extremes. \emph{Extremes}, \strong{19}, 491--515.
}
\seealso{
\code{\link[=thetafit]{thetafit()}}, \code{\link[=chifit]{chifit()}}, \code{\link[=depfit]{depfit()}}
}
\author{
\strong{Maintainer}: Thomas Lugrin \email{thomas.lugrin@alumni.epfl.ch} [copyright holder]

}
\keyword{internal}
